e2e docker fixes

+++ migrators/e2e_tests/cluster.sh
@@ -45,6 +45,11 @@ set -o errexit
 set -o nounset
 set -o pipefail
 
+export KUBE_VERBOSE=0
+export KUBE_BUILD_HYPERKUBE=n
+export KUBE_BUILD_CONFORMANCE=n
+export KUBE_BUILD_PLATFORMS="$GOOS/$GOARCH"
+
 KUBE_ROOT=$(dirname "${BASH_SOURCE[0]}")
 source "${KUBE_ROOT}/build/common.sh"
 source "${KUBE_ROOT}/build/lib/release.sh"
@@ -54,18 +59,18 @@ DEST_DIR="_output/dockerized/bin/${GOOS}/${GOARCH}"
 
 mkdir -p "${DEST_DIR}"
 
-cp "${ARTIFACTS_DIR}/kubeadm" "./${DEST_DIR}/kubeadm"
-cp "${ARTIFACTS_DIR}/kubelet" "./${DEST_DIR}/kubelet"
-cp "${ARTIFACTS_DIR}/kubectl" "./${DEST_DIR}/kubectl"
-cp "${ARTIFACTS_DIR}/kube-apiserver" "./${DEST_DIR}/kube-apiserver"
-cp "${ARTIFACTS_DIR}/kube-controller-manager" "./${DEST_DIR}/kube-controller-manager"
-cp "${ARTIFACTS_DIR}/kube-proxy" "./${DEST_DIR}/kube-proxy"
-cp "${ARTIFACTS_DIR}/kube-scheduler" "./${DEST_DIR}/kube-scheduler"
+[[ -f "./${DEST_DIR}/kubeadm" ]] || cp "${ARTIFACTS_DIR}/kubeadm" "./${DEST_DIR}/kubeadm"
+[[ -f "./${DEST_DIR}/kubelet" ]] || cp "${ARTIFACTS_DIR}/kubelet" "./${DEST_DIR}/kubelet"
+[[ -f "./${DEST_DIR}/kubectl" ]] || cp "${ARTIFACTS_DIR}/kubectl" "./${DEST_DIR}/kubectl"
+[[ -f "./${DEST_DIR}/kube-apiserver" ]] || cp "${ARTIFACTS_DIR}/kube-apiserver" "./${DEST_DIR}/kube-apiserver"
+[[ -f "./${DEST_DIR}/kube-controller-manager" ]] || cp "${ARTIFACTS_DIR}/kube-controller-manager" "./${DEST_DIR}/kube-controller-manager"
+[[ -f "./${DEST_DIR}/kube-proxy" ]] || cp "${ARTIFACTS_DIR}/kube-proxy" "./${DEST_DIR}/kube-proxy"
+[[ -f "./${DEST_DIR}/kube-scheduler" ]] || cp "${ARTIFACTS_DIR}/kube-scheduler" "./${DEST_DIR}/kube-scheduler"
 
 kube::release::build_server_images
 kind build node-image --kube-root "${KUBE_ROOT}"
 
-cp "./${DEST_DIR}/kubectl" ./kubectl
+[[ -f ./kubectl ]] || cp "./${DEST_DIR}/kubectl" ./kubectl
 
 EOF
